refactor: change MainPage to Main to get rid off the tautology
