+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.scss';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import Header from './components/Header';
-import MainPage from './pages/MainPage';
+import MainPage from './pages/Main';
 
 function App() {
   return (
